requirements.txt
Flask
numpy
tensorflow
werkzeug
Pillow
keras
opencv-python
scikit-learn

# pip install -r requirements.txt
